fix backend client and model

+++ backend/model/model.go
@@ -12,13 +12,13 @@ type AirportResponse []struct {
    ICAO             string    `json:"icao"`
    LAT              float32    `json:"lat"`
    LNG              float32    `json:"lng"`
-   ALT              float32    `json:"alt"`
+   ALT              int    `json:"alt"`
 }
 
 //TextOutput is exported,it formats the data to plain text.
 func (airport AirportResponse) TextOutput() string {
    response_out := fmt.Sprintf(
-   "UID: %s\nNAME : %s\nIATA: %s\nICAO: %s\nLAT: %f\nLNG: %f\nALT: %f\n",
+   "UID: %s\nNAME : %s\nIATA: %s\nICAO: %s\nLAT: %f\nLNG: %f\nALT: %d\n",
    airport[0].UID, airport[0].NAME, airport[0].IATA, airport[0].ICAO, airport[0].LAT, airport[0].LNG, airport[0].ALT)
    return response_out
 }
